Fixed the install and run scripts, minor edits
I changed the scripts to use a set Python executable, since for my
distro the default python2 executable is 2.7 (but python27 is still the
default). Also removed the ping for the bitmessage server as it caused a
crash almost every time. Lastly added to the gitignore the bitcoin and
bitmessage folders.